Add a basic pathfinding debug renderer
Doesn't include all the wanted information, but it does render the
current path.

This is done using mixins.  It could be done even further (or more accurately) by expanding the mixins, but I haven't done that (I don't know the entire pathfinding system and I don't know what correct values for the open and closed sets would be).